+++ modules/images/src/index.js
@@ -4,10 +4,7 @@ export {default as ImageWriter} from './image-writer';
 export {loadImage} from './lib/parse-image';
 
 // UTILS
-export {isImage, getImageMetadata} from './lib/get-image-metadata';
+export {isImage, getImageMetadata, getImageMIMEType, getImageSize} from './lib/get-image-metadata';
 
 // Experimental
 export {decodeImage} from './lib/image-utils-browser';
-
-// Deprecated
-export {getImageSize} from './lib/get-image-metadata';